requirements.txt
slack
python-dotenv
ezsheets
pandas
fuzz